requirements.txt
flask==3.0.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0
requests==2.31.0
icalendar==5.0.12
python-dateutil==2.9.0
pytz==2024.1
